Update interception.js
Improvement:
It does recovers deleted message but when I load the whatsapp, it treats them as a normal message without red color.
Let me be clear:
I was offline on whatsapp web, I sent myself a message from my other phone and when I turned on whatsapp web, i could see the message but could not tell whether that was a normal message or deleted (Blue ticks are off). So make a feature like that so we could know that the recovered message was a deleted message.